Add in diagnostics page
